Windows MSVC fixes

+++ include/bvh.h
@@ -1,9 +1,7 @@
 #ifndef BVH_H
 #define BVH_H
 
-#include <glad/glad.h>
 #include <glm/glm.hpp>
-#include <limits>
 #include <vector>
 #include "object.h"
 
@@ -27,7 +25,7 @@ struct BVH
 
     BVH(std::vector<GPUTriangle> &triangles);
 
-    void split(const uint32_t nodeIndex, const int depth = 0);
+    void split(const uint32_t nodeIndex, const size_t depth = 0);
     void growToInclude(GPUNode &node, const glm::vec3 point);
     void growToInclude(GPUNode &node, const GPUTriangle &triangle);
 };
